requirements.txt
torch
transformers
sentencepiece
protobuf
pillow
numpy
matplotlib
opencv-python
easyocr
pyzbar